Really fix uri().
What uri() returns has never been properly documented until now. 
Even Andy was confused, thinking it returned a URI object when 
it has (lately) been returning a string. 

In Mech 1.04, it sometimes returned a string and sometimes a URI object,
depending on when you called uri(). 

Sometime after that, it was changed to always return a regular string, 
and about 20 nearly identical (!!) tests were added to the test suite to
insure this was always case. 

Unfortunately, the /wrong/ URI was being returned after a redirect, causing
three bugs to be filed about the issue, and the superior solution was
overlooked: a URI object can returned which happens to stringify to a string. 

So we get the best of both worlds and backwards compatibility. 


git-svn-id: http://svn.perl.org/modules/www-mechanize/trunk@6645 50811bd7-b8ce-0310-adc1-d9db26280581